requirements.txt
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
python-dotenv==0.21.0
pyziptax==1.2.1
requests==2.28.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
